@@ -67,7 +80,7 @@
     "tailwindcss": "^3.0.13",
     "ts-loader": "^9.2.6",
     "ts-node": "^10.4.0",
-    "typescript": "^4.5.4",
+    "typescript": "^4.7.4",
     "url-loader": "^4.1.1",
     "webpack": "^5.65.0",
     "webpack-bundle-analyzer": "^4.5.0",
