@@ -5,13 +5,15 @@
 
 #include "as_cli.hpp"
 
-auto Friend::generate_key(unique_ptr<asphrdaemon::Daemon::Stub>& stub)
+// TODO: Due to API change, this file needs to be rewritten
+
+/**auto Friend::generate_key(unique_ptr<asphrdaemon::Daemon::Stub>& stub)
     -> asphr::StatusOr<string> {
   grpc::ClientContext context;
   asphrdaemon::GenerateFriendKeyRequest request;
   asphrdaemon::GenerateFriendKeyResponse response;
 
-  request.set_name(name_);
+  request.set_unique_name(name_);
 
   grpc::Status status = stub->GenerateFriendKey(&context, request, &response);
 
@@ -28,7 +30,7 @@ auto Friend::add(unique_ptr<asphrdaemon::Daemon::Stub>& stub, const string& key)
   asphrdaemon::AddFriendRequest request;
   asphrdaemon::AddFriendResponse response;
 
-  request.set_name(name_);
+  request.set_unique_name(name_);
   // REMOVE
   cout << "name: " << name_ << endl;
   request.set_key(key);
@@ -42,4 +44,4 @@ auto Friend::add(unique_ptr<asphrdaemon::Daemon::Stub>& stub, const string& key)
   return absl::OkStatus();
 }
 
-bool Friend::complete() const { return !name_is_empty(); }
+bool Friend::complete() const { return !name_is_empty(); }**/
