@@ -42,8 +42,9 @@ auto Profile::get_friends(unique_ptr<asphrdaemon::Daemon::Stub>& stub)
   Friend::FriendMap friends_map;
 
   for (const auto& friend_info : reply.friend_infos()) {
-    Friend friend_struct(friend_info.name());
-    friends_map.insert(std::make_pair(friend_info.name(), friend_struct));
+    Friend friend_struct(friend_info.unique_name());
+    friends_map.insert(
+        std::make_pair(friend_info.unique_name(), friend_struct));
   }
 
   return friends_map;
