@@ -10,10 +10,18 @@ import * as grpc from "@grpc/grpc-js";
 
 interface IDaemonService extends grpc.ServiceDefinition<grpc.UntypedServiceImplementation> {
   registerUser: grpc.MethodDefinition<schema_daemon_pb.RegisterUserRequest, schema_daemon_pb.RegisterUserResponse>;
+  getMyPublicID: grpc.MethodDefinition<schema_daemon_pb.GetMyPublicIDRequest, schema_daemon_pb.GetMyPublicIDResponse>;
+  isValidPublicID: grpc.MethodDefinition<schema_daemon_pb.IsValidPublicIDRequest, schema_daemon_pb.IsValidPublicIDResponse>;
   getFriendList: grpc.MethodDefinition<schema_daemon_pb.GetFriendListRequest, schema_daemon_pb.GetFriendListResponse>;
-  generateFriendKey: grpc.MethodDefinition<schema_daemon_pb.GenerateFriendKeyRequest, schema_daemon_pb.GenerateFriendKeyResponse>;
-  addFriend: grpc.MethodDefinition<schema_daemon_pb.AddFriendRequest, schema_daemon_pb.AddFriendResponse>;
   removeFriend: grpc.MethodDefinition<schema_daemon_pb.RemoveFriendRequest, schema_daemon_pb.RemoveFriendResponse>;
+  addSyncFriend: grpc.MethodDefinition<schema_daemon_pb.AddSyncFriendRequest, schema_daemon_pb.AddSyncFriendResponse>;
+  addAsyncFriend: grpc.MethodDefinition<schema_daemon_pb.AddAsyncFriendRequest, schema_daemon_pb.AddAsyncFriendResponse>;
+  getOutgoingSyncInvitations: grpc.MethodDefinition<schema_daemon_pb.GetOutgoingSyncInvitationsRequest, schema_daemon_pb.GetOutgoingSyncInvitationsResponse>;
+  getOutgoingAsyncInvitations: grpc.MethodDefinition<schema_daemon_pb.GetOutgoingAsyncInvitationsRequest, schema_daemon_pb.GetOutgoingAsyncInvitationsResponse>;
+  getIncomingAsyncInvitations: grpc.MethodDefinition<schema_daemon_pb.GetIncomingAsyncInvitationsRequest, schema_daemon_pb.GetIncomingAsyncInvitationsResponse>;
+  acceptAsyncInvitation: grpc.MethodDefinition<schema_daemon_pb.AcceptAsyncInvitationRequest, schema_daemon_pb.AcceptAsyncInvitationResponse>;
+  rejectAsyncInvitation: grpc.MethodDefinition<schema_daemon_pb.RejectAsyncInvitationRequest, schema_daemon_pb.RejectAsyncInvitationResponse>;
+  cancelAsyncInvitation: grpc.MethodDefinition<schema_daemon_pb.CancelAsyncInvitationRequest, schema_daemon_pb.CancelAsyncInvitationResponse>;
   sendMessage: grpc.MethodDefinition<schema_daemon_pb.SendMessageRequest, schema_daemon_pb.SendMessageResponse>;
   getMessages: grpc.MethodDefinition<schema_daemon_pb.GetMessagesRequest, schema_daemon_pb.GetMessagesResponse>;
   getMessagesStreamed: grpc.MethodDefinition<schema_daemon_pb.GetMessagesRequest, schema_daemon_pb.GetMessagesResponse>;
@@ -30,10 +38,18 @@ export const DaemonService: IDaemonService;
 
 export interface IDaemonServer extends grpc.UntypedServiceImplementation {
   registerUser: grpc.handleUnaryCall<schema_daemon_pb.RegisterUserRequest, schema_daemon_pb.RegisterUserResponse>;
+  getMyPublicID: grpc.handleUnaryCall<schema_daemon_pb.GetMyPublicIDRequest, schema_daemon_pb.GetMyPublicIDResponse>;
+  isValidPublicID: grpc.handleUnaryCall<schema_daemon_pb.IsValidPublicIDRequest, schema_daemon_pb.IsValidPublicIDResponse>;
   getFriendList: grpc.handleUnaryCall<schema_daemon_pb.GetFriendListRequest, schema_daemon_pb.GetFriendListResponse>;
-  generateFriendKey: grpc.handleUnaryCall<schema_daemon_pb.GenerateFriendKeyRequest, schema_daemon_pb.GenerateFriendKeyResponse>;
-  addFriend: grpc.handleUnaryCall<schema_daemon_pb.AddFriendRequest, schema_daemon_pb.AddFriendResponse>;
   removeFriend: grpc.handleUnaryCall<schema_daemon_pb.RemoveFriendRequest, schema_daemon_pb.RemoveFriendResponse>;
+  addSyncFriend: grpc.handleUnaryCall<schema_daemon_pb.AddSyncFriendRequest, schema_daemon_pb.AddSyncFriendResponse>;
+  addAsyncFriend: grpc.handleUnaryCall<schema_daemon_pb.AddAsyncFriendRequest, schema_daemon_pb.AddAsyncFriendResponse>;
+  getOutgoingSyncInvitations: grpc.handleUnaryCall<schema_daemon_pb.GetOutgoingSyncInvitationsRequest, schema_daemon_pb.GetOutgoingSyncInvitationsResponse>;
+  getOutgoingAsyncInvitations: grpc.handleUnaryCall<schema_daemon_pb.GetOutgoingAsyncInvitationsRequest, schema_daemon_pb.GetOutgoingAsyncInvitationsResponse>;
+  getIncomingAsyncInvitations: grpc.handleUnaryCall<schema_daemon_pb.GetIncomingAsyncInvitationsRequest, schema_daemon_pb.GetIncomingAsyncInvitationsResponse>;
+  acceptAsyncInvitation: grpc.handleUnaryCall<schema_daemon_pb.AcceptAsyncInvitationRequest, schema_daemon_pb.AcceptAsyncInvitationResponse>;
+  rejectAsyncInvitation: grpc.handleUnaryCall<schema_daemon_pb.RejectAsyncInvitationRequest, schema_daemon_pb.RejectAsyncInvitationResponse>;
+  cancelAsyncInvitation: grpc.handleUnaryCall<schema_daemon_pb.CancelAsyncInvitationRequest, schema_daemon_pb.CancelAsyncInvitationResponse>;
   sendMessage: grpc.handleUnaryCall<schema_daemon_pb.SendMessageRequest, schema_daemon_pb.SendMessageResponse>;
   getMessages: grpc.handleUnaryCall<schema_daemon_pb.GetMessagesRequest, schema_daemon_pb.GetMessagesResponse>;
   getMessagesStreamed: grpc.handleServerStreamingCall<schema_daemon_pb.GetMessagesRequest, schema_daemon_pb.GetMessagesResponse>;
@@ -51,18 +67,42 @@ export class DaemonClient extends grpc.Client {
   registerUser(argument: schema_daemon_pb.RegisterUserRequest, callback: grpc.requestCallback<schema_daemon_pb.RegisterUserResponse>): grpc.ClientUnaryCall;
   registerUser(argument: schema_daemon_pb.RegisterUserRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.RegisterUserResponse>): grpc.ClientUnaryCall;
   registerUser(argument: schema_daemon_pb.RegisterUserRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.RegisterUserResponse>): grpc.ClientUnaryCall;
+  getMyPublicID(argument: schema_daemon_pb.GetMyPublicIDRequest, callback: grpc.requestCallback<schema_daemon_pb.GetMyPublicIDResponse>): grpc.ClientUnaryCall;
+  getMyPublicID(argument: schema_daemon_pb.GetMyPublicIDRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GetMyPublicIDResponse>): grpc.ClientUnaryCall;
+  getMyPublicID(argument: schema_daemon_pb.GetMyPublicIDRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GetMyPublicIDResponse>): grpc.ClientUnaryCall;
+  isValidPublicID(argument: schema_daemon_pb.IsValidPublicIDRequest, callback: grpc.requestCallback<schema_daemon_pb.IsValidPublicIDResponse>): grpc.ClientUnaryCall;
+  isValidPublicID(argument: schema_daemon_pb.IsValidPublicIDRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.IsValidPublicIDResponse>): grpc.ClientUnaryCall;
+  isValidPublicID(argument: schema_daemon_pb.IsValidPublicIDRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.IsValidPublicIDResponse>): grpc.ClientUnaryCall;
   getFriendList(argument: schema_daemon_pb.GetFriendListRequest, callback: grpc.requestCallback<schema_daemon_pb.GetFriendListResponse>): grpc.ClientUnaryCall;
   getFriendList(argument: schema_daemon_pb.GetFriendListRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GetFriendListResponse>): grpc.ClientUnaryCall;
   getFriendList(argument: schema_daemon_pb.GetFriendListRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GetFriendListResponse>): grpc.ClientUnaryCall;
-  generateFriendKey(argument: schema_daemon_pb.GenerateFriendKeyRequest, callback: grpc.requestCallback<schema_daemon_pb.GenerateFriendKeyResponse>): grpc.ClientUnaryCall;
-  generateFriendKey(argument: schema_daemon_pb.GenerateFriendKeyRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GenerateFriendKeyResponse>): grpc.ClientUnaryCall;
-  generateFriendKey(argument: schema_daemon_pb.GenerateFriendKeyRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GenerateFriendKeyResponse>): grpc.ClientUnaryCall;
-  addFriend(argument: schema_daemon_pb.AddFriendRequest, callback: grpc.requestCallback<schema_daemon_pb.AddFriendResponse>): grpc.ClientUnaryCall;
-  addFriend(argument: schema_daemon_pb.AddFriendRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.AddFriendResponse>): grpc.ClientUnaryCall;
-  addFriend(argument: schema_daemon_pb.AddFriendRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.AddFriendResponse>): grpc.ClientUnaryCall;
   removeFriend(argument: schema_daemon_pb.RemoveFriendRequest, callback: grpc.requestCallback<schema_daemon_pb.RemoveFriendResponse>): grpc.ClientUnaryCall;
   removeFriend(argument: schema_daemon_pb.RemoveFriendRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.RemoveFriendResponse>): grpc.ClientUnaryCall;
   removeFriend(argument: schema_daemon_pb.RemoveFriendRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.RemoveFriendResponse>): grpc.ClientUnaryCall;
+  addSyncFriend(argument: schema_daemon_pb.AddSyncFriendRequest, callback: grpc.requestCallback<schema_daemon_pb.AddSyncFriendResponse>): grpc.ClientUnaryCall;
+  addSyncFriend(argument: schema_daemon_pb.AddSyncFriendRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.AddSyncFriendResponse>): grpc.ClientUnaryCall;
+  addSyncFriend(argument: schema_daemon_pb.AddSyncFriendRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.AddSyncFriendResponse>): grpc.ClientUnaryCall;
+  addAsyncFriend(argument: schema_daemon_pb.AddAsyncFriendRequest, callback: grpc.requestCallback<schema_daemon_pb.AddAsyncFriendResponse>): grpc.ClientUnaryCall;
+  addAsyncFriend(argument: schema_daemon_pb.AddAsyncFriendRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.AddAsyncFriendResponse>): grpc.ClientUnaryCall;
+  addAsyncFriend(argument: schema_daemon_pb.AddAsyncFriendRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.AddAsyncFriendResponse>): grpc.ClientUnaryCall;
+  getOutgoingSyncInvitations(argument: schema_daemon_pb.GetOutgoingSyncInvitationsRequest, callback: grpc.requestCallback<schema_daemon_pb.GetOutgoingSyncInvitationsResponse>): grpc.ClientUnaryCall;
+  getOutgoingSyncInvitations(argument: schema_daemon_pb.GetOutgoingSyncInvitationsRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GetOutgoingSyncInvitationsResponse>): grpc.ClientUnaryCall;
+  getOutgoingSyncInvitations(argument: schema_daemon_pb.GetOutgoingSyncInvitationsRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GetOutgoingSyncInvitationsResponse>): grpc.ClientUnaryCall;
+  getOutgoingAsyncInvitations(argument: schema_daemon_pb.GetOutgoingAsyncInvitationsRequest, callback: grpc.requestCallback<schema_daemon_pb.GetOutgoingAsyncInvitationsResponse>): grpc.ClientUnaryCall;
+  getOutgoingAsyncInvitations(argument: schema_daemon_pb.GetOutgoingAsyncInvitationsRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GetOutgoingAsyncInvitationsResponse>): grpc.ClientUnaryCall;
+  getOutgoingAsyncInvitations(argument: schema_daemon_pb.GetOutgoingAsyncInvitationsRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GetOutgoingAsyncInvitationsResponse>): grpc.ClientUnaryCall;
+  getIncomingAsyncInvitations(argument: schema_daemon_pb.GetIncomingAsyncInvitationsRequest, callback: grpc.requestCallback<schema_daemon_pb.GetIncomingAsyncInvitationsResponse>): grpc.ClientUnaryCall;
+  getIncomingAsyncInvitations(argument: schema_daemon_pb.GetIncomingAsyncInvitationsRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GetIncomingAsyncInvitationsResponse>): grpc.ClientUnaryCall;
+  getIncomingAsyncInvitations(argument: schema_daemon_pb.GetIncomingAsyncInvitationsRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.GetIncomingAsyncInvitationsResponse>): grpc.ClientUnaryCall;
+  acceptAsyncInvitation(argument: schema_daemon_pb.AcceptAsyncInvitationRequest, callback: grpc.requestCallback<schema_daemon_pb.AcceptAsyncInvitationResponse>): grpc.ClientUnaryCall;
+  acceptAsyncInvitation(argument: schema_daemon_pb.AcceptAsyncInvitationRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.AcceptAsyncInvitationResponse>): grpc.ClientUnaryCall;
+  acceptAsyncInvitation(argument: schema_daemon_pb.AcceptAsyncInvitationRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.AcceptAsyncInvitationResponse>): grpc.ClientUnaryCall;
+  rejectAsyncInvitation(argument: schema_daemon_pb.RejectAsyncInvitationRequest, callback: grpc.requestCallback<schema_daemon_pb.RejectAsyncInvitationResponse>): grpc.ClientUnaryCall;
+  rejectAsyncInvitation(argument: schema_daemon_pb.RejectAsyncInvitationRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.RejectAsyncInvitationResponse>): grpc.ClientUnaryCall;
+  rejectAsyncInvitation(argument: schema_daemon_pb.RejectAsyncInvitationRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.RejectAsyncInvitationResponse>): grpc.ClientUnaryCall;
+  cancelAsyncInvitation(argument: schema_daemon_pb.CancelAsyncInvitationRequest, callback: grpc.requestCallback<schema_daemon_pb.CancelAsyncInvitationResponse>): grpc.ClientUnaryCall;
+  cancelAsyncInvitation(argument: schema_daemon_pb.CancelAsyncInvitationRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.CancelAsyncInvitationResponse>): grpc.ClientUnaryCall;
+  cancelAsyncInvitation(argument: schema_daemon_pb.CancelAsyncInvitationRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.CancelAsyncInvitationResponse>): grpc.ClientUnaryCall;
   sendMessage(argument: schema_daemon_pb.SendMessageRequest, callback: grpc.requestCallback<schema_daemon_pb.SendMessageResponse>): grpc.ClientUnaryCall;
   sendMessage(argument: schema_daemon_pb.SendMessageRequest, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.SendMessageResponse>): grpc.ClientUnaryCall;
   sendMessage(argument: schema_daemon_pb.SendMessageRequest, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<schema_daemon_pb.SendMessageResponse>): grpc.ClientUnaryCall;
