@@ -3,10 +3,15 @@
 // SPDX-License-Identifier: GPL-3.0-only
 //
 
-/* eslint global-require: off, no-console: off, promise/always-return: off */
-
 import path from "path";
-import { app, BrowserWindow, session, shell, Notification } from "electron";
+import {
+  app,
+  dialog,
+  BrowserWindow,
+  session,
+  shell,
+  Notification,
+} from "electron";
 import MenuBuilder from "./menu";
 import { resolveHtmlPath } from "./util";
 import { autoUpdater } from "electron-updater";
@@ -14,89 +19,133 @@ import { promisify } from "util";
 import { exec as execNonPromisified } from "child_process";
 const exec = promisify(execNonPromisified);
 
-import {
-  getDaemonClient,
-  convertProtobufIncomingMessageToTypedMessage,
-  truncate,
-} from "./daemon";
-import daemonM from "../daemon/schema/daemon_pb";
+import { truncate, DaemonImpl } from "./daemon";
+import { IncomingMessage } from "../types";
+import * as daemon_pb from "../daemon/schema/daemon_pb";
 import { PLIST_CONTENTS, PLIST_PATH, RELEASE_COMMIT_HASH } from "./constants";
 import { exit } from "process";
 import fs from "fs";
-const daemonClient = getDaemonClient();
+import { Console } from "console";
+
+const daemon = new DaemonImpl();
 
 const isDevelopment =
-  process.env.NODE_ENV === "development" || process.env.DEBUG_PROD === "true";
+  process.env["NODE_ENV"] === "development" ||
+  process.env["DEBUG_PROD"] === "true";
 
-if (process.env.NODE_ENV === "production") {
+if (process.env["NODE_ENV"] === "production") {
   const sourceMapSupport = require("source-map-support");
   sourceMapSupport.install();
 }
 
-const startDaemonIfNeeded = async (pkgPath: string) => {
-  const request = new daemonM.GetStatusRequest();
-  const getStatus = promisify(daemonClient.getStatus).bind(daemonClient);
+if (process.defaultApp) {
+  if (process.argv.length >= 2 && process.argv[1] != null) {
+    app.setAsDefaultProtocolClient("anysphere", process.execPath, [
+      path.resolve(process.argv[1]),
+    ]);
+  }
+} else {
+  app.setAsDefaultProtocolClient("anysphere");
+}
+
+function setupLogger(): void {
+  let logPath = "";
+  if (process.env["XDG_CACHE_HOME"] !== undefined) {
+    logPath = path.join(process.env["XDG_CACHE_HOME"], "anysphere", "logs");
+  } else if (process.env["HOME"] !== undefined) {
+    logPath = path.join(process.env["HOME"], ".anysphere", "cache", "logs");
+  } else {
+    process.stderr.write(
+      "$HOME or $XDG_CACHE_HOME not set! Cannot create log path, aborting :("
+    );
+    exit(1);
+  }
+  app.setAppLogsPath(logPath);
+
+  // in debug mode, use console.log. in production, use the log file.
+  if (isDevelopment) {
+    global.logger = console;
+  } else {
+    global.logger = new Console({
+      stdout: fs.createWriteStream(path.join(logPath, "anysphere-gui.log")),
+      stderr: fs.createWriteStream(path.join(logPath, "anysphere-gui.err")),
+    });
+  }
+}
+
+async function startDaemonIfNeeded(pkgPath: string): Promise<void> {
   try {
-    const response = (await getStatus(request)) as daemonM.GetStatusResponse;
+    const daemonStatus = await daemon.getStatus({});
     // if release hash is wrong, we need to restart!
-    if (response.getReleaseHash() !== RELEASE_COMMIT_HASH) {
+    if (daemonStatus.releaseHash !== RELEASE_COMMIT_HASH) {
       throw new Error("incorrect release hash");
     }
     // daemon is running, correct version, nothing to do
+    logger.log("Daemon is running, with the correct version!");
     return;
   } catch (e) {
     // if development, we don't want to start the daemon (want to do it manually)
     if (isDevelopment) {
-      process.stdout.write(
+      logger.log(
         `Daemon is either not running or running the wrong version. Please start it; we're not doing anything because we're in DEV mode. Error: ${e}.`
       );
       return;
+    } else {
+      logger.log(
+        `Daemon is either not running or running the wrong version. Error: ${e}.`
+      );
     }
 
+    // We do not copy the CLI here, because installing the CLI is a pain.
+    // Particularly, we would need to modify the user's PATH.
+    // TODO: expose a cmd-K command to install the CLI. It presents the user with a
+    // copy-able short command to run in the terminal.
+    //
     // first copy the CLI
-    const cliPath = path.join(pkgPath, "bin", "anysphere");
-    // ln -sf link it!
-    const mkdir = await exec(`mkdir -p /usr/local/bin`);
-    if (mkdir.stderr) {
-      process.stderr.write(mkdir.stderr);
-    }
-    const clilink = await exec(`ln -sf ${cliPath} /usr/local/bin/anysphere`);
-    if (clilink.stderr) {
-      process.stderr.write(clilink.stderr);
-    }
+    // const cliPath = path.join(pkgPath, "bin", "anysphere");
+    // // ln -sf link it!
+    // // TODO(arvid): just add to PATH instead, because not everyone has /usr/local/bin in their PATH
+    // const mkdir = await exec(`mkdir -p /usr/local/bin`);
+    // if (mkdir.stderr) {
+    //   logger.error(mkdir.stderr);
+    // }
+    // const clilink = await exec(`ln -sf ${cliPath} /usr/local/bin/anysphere`);
+    // if (clilink.stderr) {
+    //   logger.error(clilink.stderr);
+    // }
+    // logger.log("Successfully linked the CLI.");
 
     // TODO(arvid): handle windows and linux too
     // possible problem, so let's start the daemon!
     // unload the plist if it exists.
-    const plist_path = PLIST_PATH();
+    const plistPath = PLIST_PATH();
+    // 0: create the directory
+    const mkdirPlist = await exec(`mkdir -p ${path.dirname(plistPath)}`);
+    if (mkdirPlist.stderr) {
+      logger.error(mkdirPlist.stderr);
+    }
+    logger.log("Successfully created the plist directory.");
     // 1: unload plist
-    await exec("launchctl unload " + plist_path); // we don't care if it fails or not!
-    let logPath = "";
-    if (process.env.XDG_CACHE_HOME) {
-      logPath = path.join(process.env.XDG_CACHE_HOME, "anysphere", "logs");
-    } else if (process.env.HOME) {
-      logPath = path.join(process.env.HOME, ".anysphere", "cache", "logs");
-    } else {
-      process.stderr.write(
-        "$HOME or $XDG_CACHE_HOME not set! Cannot create daemon, aborting :("
-      );
-      exit(1);
-    }
+    await exec("launchctl unload " + plistPath); // we don't care if it fails or not!
+    const logPath = app.getPath("logs");
     const contents = PLIST_CONTENTS(pkgPath, logPath);
+    logger.log("Successfully unloaded the plist.");
     // 2: write plist
-    await fs.promises.writeFile(plist_path, contents);
+    await fs.promises.writeFile(plistPath, contents);
+    logger.log("Successfully wrote the new plist.");
     // 3: load plist
-    const response = await exec("launchctl load " + plist_path);
+    const response = await exec("launchctl load " + plistPath);
     if (response.stderr) {
-      process.stderr.write(response.stderr);
+      logger.error(response.stderr);
       exit(1);
     }
+    logger.log("Successfully loaded the new plist.");
   }
-};
+}
 
-const installExtensions = async () => {
+async function installExtensions(): Promise<void> {
   const installer = require("electron-devtools-installer");
-  const forceDownload = !!process.env.UPGRADE_EXTENSIONS;
+  const forceDownload = !!process.env["UPGRADE_EXTENSIONS"];
   const extensions = ["REACT_DEVELOPER_TOOLS"];
 
   return installer
@@ -104,10 +153,10 @@ const installExtensions = async () => {
       extensions.map((name) => installer[name]),
       forceDownload
     )
-    .catch(console.log);
-};
+    .catch(logger.log);
+}
 
-const createWindow = async () => {
+const createWindow = async (publicID?: string): Promise<void> => {
   if (isDevelopment) {
     await installExtensions();
   }
@@ -136,13 +185,15 @@ const createWindow = async () => {
     },
   });
 
-  mainWindow.loadURL(resolveHtmlPath("index.html"));
+  let loadURL = resolveHtmlPath("index.html");
+  if (publicID != null) {
+    loadURL = resolveHtmlPath(`index.html?publicID=${publicID}`);
+  }
+
+  mainWindow.loadURL(loadURL);
 
   mainWindow.on("ready-to-show", () => {
-    if (!mainWindow) {
-      throw new Error('"mainWindow" is not defined');
-    }
-    if (process.env.START_MINIMIZED) {
+    if (process.env["START_MINIMIZED"] === "true") {
       mainWindow.minimize();
     } else {
       mainWindow.show();
@@ -154,12 +205,17 @@ const createWindow = async () => {
 
   // Don't allow ANY requests to any origin! This means that the app will
   // only not be able to communicate with the internet at all, which is PERFECT.
+  session.defaultSession.enableNetworkEmulation({
+    offline: true,
+  });
   session.defaultSession.webRequest.onHeadersReceived((details, callback) => {
     callback({
       responseHeaders: {
         ...details.responseHeaders,
+        // eslint-disable-next-line @typescript-eslint/naming-convention
         "Content-Security-Policy": [
-          "default-src 'self' style-src 'self' 'unsafe-inline'",
+          "default-src 'self'; style-src 'self' 'unsafe-inline'", // required for tailwind
+          // "default-src 'self'",
         ],
       },
     });
@@ -178,57 +234,36 @@ app.on("window-all-closed", () => {
   }
 });
 
-function registerForNotifications() {
-  const request = new daemonM.GetMessagesRequest();
-  request.setFilter(daemonM.GetMessagesRequest.Filter.NEW);
-  var call = daemonClient.getMessagesStreamed(request);
-
+function registerForNotifications(): () => void {
   let firstTime = true;
-
-  call.on("data", function (r) {
-    if (firstTime) {
-      firstTime = false;
-      return;
-    }
-    try {
-      const lm = r.getMessagesList();
-      const l = lm.map(convertProtobufIncomingMessageToTypedMessage);
-      // notify!!!
-      for (const m of l) {
+  const cancel = daemon.getMessagesStreamed(
+    { filter: daemon_pb.GetMessagesRequest.Filter.NEW },
+    (messages: IncomingMessage[]) => {
+      if (firstTime) {
+        firstTime = false;
+        return;
+      }
+      for (const m of messages) {
         const notification = new Notification({
-          title: m.from,
+          title: m.fromDisplayName,
           body: truncate(m.message, 50),
         });
         notification.show();
       }
-    } catch (e) {
-      console.log(`error in getNewMessagesStreamed: ${e}`);
     }
-    console.log("got all messages streamed", r);
-  });
-  call.on("end", function () {
-    // The server has finished sending
-    // TODO(arvid): resubscribe?
-    console.log("getNewMessagesStreamed end");
-  });
-  call.on("error", function (e) {
-    // An error has occurred and the stream has been closed.
-    // TODO(arvid): resubscribe?
-    console.log("getNewMessagesStreamed error", e);
-  });
-  call.on("status", function (status) {
-    // process status
-    console.log("getNewMessagesStreamed status", status);
-  });
+  );
+
   return () => {
     console.log("cancelling notifications!");
-    call.cancel();
+    cancel();
   };
 }
 
 app
   .whenReady()
   .then(() => {
+    setupLogger();
+
     autoUpdater.checkForUpdatesAndNotify();
 
     startDaemonIfNeeded(path.dirname(app.getAppPath()));
@@ -256,4 +291,16 @@ app.on("web-contents-created", (event, contents) => {
     event.preventDefault();
     shell.openExternal(navigationUrl);
   });
+});
+
+// Handle the protocol. In this case, we choose to show an Error Box.
+app.on("open-url", (event, url) => {
+  const publicID = url.split("/").pop();
+  if (BrowserWindow.getAllWindows().length === 0) {
+    createWindow(publicID);
+  } else {
+    BrowserWindow.getAllWindows()[0].loadURL(
+      resolveHtmlPath("index.html?publicID=" + publicID)
+    );
+  }
 });
