@@ -4,132 +4,172 @@
 //
 
 import * as React from "react";
-import { Message } from "../../types";
+import { IncomingMessage, OutgoingMessage } from "../../types";
+import * as daemon_pb from "../../daemon/schema/daemon_pb";
 import { truncate, formatTime } from "../utils";
+import { BackgroundCircles } from "./BackgroundCircles/BackgroundCircles";
 import { SelectableList } from "./SelectableList";
+import { protobufDateToDate } from "../../types";
 
-function MessageBlurb({
+function IncomingMessageBlurb({
   message,
   active,
 }: {
-  message: Message;
+  message: IncomingMessage;
   active: boolean;
-}) {
-  let timestamp_string = "";
-  try {
-    timestamp_string = formatTime(message.timestamp);
-  } catch {}
+}): JSX.Element {
+  let timestampString = "";
+
+  if (message.deliveredAt === undefined) {
+    timestampString = "Not yet delivered.";
+    console.error("Message has no deliveredAt");
+  } else {
+    timestampString = formatTime(protobufDateToDate(message.deliveredAt));
+  }
 
   return (
     <div
       className={`${
-        active ? "bg-asbeige border-asbrown-100" : "bg-white border-white"
-      } border-l-4 px-4 py-4 rounded-sm my-2`}
+        active ? "border-asbrown-100 bg-asbeige2" : "border-white bg-white"
+      } my-2 rounded-sm border-l-4 px-4 py-4`}
     >
       <div className="flex flex-row gap-5">
-        <div className="text-asbrown-dark text-sm">
-          {message.type === "incoming" ? message.from : `To: ${message.to}`}
+        <div className="text-sm text-asbrown-dark">
+          {message.fromDisplayName}
         </div>
-        <div className="text-asbrown-300 text-sm">
+        <div className="text-sm text-asbrown-300">
           {truncate(message.message, 65)}
         </div>
         <div className="flex-1"></div>
-        <div className="text-asbrown-200 text-sm">{timestamp_string}</div>
+        <div className="text-sm text-asbrown-200">{timestampString}</div>
       </div>
     </div>
   );
 }
 
-function NoMessages({ explanation }: { explanation: string }) {
+function OutgoingMessageBlurb({
+  message,
+  active,
+}: {
+  message: OutgoingMessage;
+  active: boolean;
+}): JSX.Element {
+  let timestampString = "";
+
+  if (message.sentAt === undefined) {
+    timestampString = "Not yet sent.";
+    console.error("Message has no sentAt");
+  } else {
+    timestampString = formatTime(protobufDateToDate(message.sentAt));
+  }
+
   return (
-    <div className="grid h-full pt-48">
-      <div className="place-self-center text-asbrown-200 text-xs unselectable">
-        {explanation}
+    <div
+      className={`${
+        active ? "border-asbrown-100 bg-asbeige" : "border-white bg-white"
+      } my-2 rounded-sm border-l-4 px-4 py-4`}
+    >
+      <div className="flex flex-row gap-5">
+        <div className="text-sm text-asbrown-dark">{`To: ${message.toFriendsList
+          .map((x) => x.displayName)
+          .join(", ")}`}</div>
+        <div className="text-sm text-asbrown-300">
+          {truncate(message.message, 65)}
+        </div>
+        <div className="flex-1"></div>
+        <div className="text-sm text-asbrown-200">{timestampString}</div>
       </div>
     </div>
   );
 }
 
-function MessageList(props: {
-  messages: string;
-  readCallback: (message: Message) => void;
-}) {
-  const [messages, setMessages] = React.useState<Message[]>([]);
+export function IncomingMessageList({
+  type,
+  readCallback,
+}: {
+  type: "new" | "all";
+  readCallback: (message: IncomingMessage) => void;
+}): JSX.Element {
+  const [messages, setMessages] = React.useState<IncomingMessage[]>([]);
 
   React.useEffect(() => {
-    if (props.messages === "new") {
+    if (type === "new") {
       setMessages([]);
-      let cancel = window.getNewMessagesStreamed((messages: Message[]) => {
-        setMessages((prev: Message[]) => {
-          // merge new messages with old messages, and sort them by timestamp
-          let new_messages = messages.concat(prev);
-          new_messages.sort((a, b) => {
-            // sort based on timestamp
-            if (a.timestamp > b.timestamp) {
-              return -1;
-            } else if (a.timestamp < b.timestamp) {
-              return 1;
-            } else {
-              return 0;
-            }
+      const cancel = window.getMessagesStreamed(
+        { filter: daemon_pb.GetMessagesRequest.Filter.NEW },
+        (messages: IncomingMessage[]) => {
+          setMessages((prev: IncomingMessage[]) => {
+            // merge new messages with old messages, and sort them by timestamp
+            let newMessages = messages.concat(prev);
+            newMessages = newMessages.sort((a, b) => {
+              // sort based on timestamp
+              const aTime = a.deliveredAt ?? new Date();
+              const bTime = b.deliveredAt ?? new Date();
+              if (aTime > bTime) {
+                return -1;
+              } else if (aTime < bTime) {
+                return 1;
+              } else {
+                return 0;
+              }
+            });
+            return newMessages;
           });
-          return new_messages;
-        });
-      });
-      return cancel;
-    } else if (props.messages === "all") {
-      setMessages([]);
-      let cancel = window.getAllMessagesStreamed((messages: Message[]) => {
-        setMessages((prev: Message[]) => {
-          // merge new messages with old messages, and sort them by timestamp
-          let new_messages = messages.concat(prev);
-          new_messages.sort((a, b) => {
-            // sort based on timestamp
-            if (a.timestamp > b.timestamp) {
-              return -1;
-            } else if (a.timestamp < b.timestamp) {
-              return 1;
-            } else {
-              return 0;
-            }
-          });
-          return new_messages;
-        });
-      });
-      return cancel;
-    } else if (props.messages === "outbox") {
-      window.getOutboxMessages().then((messages: Message[]) => {
-        setMessages(messages);
-      });
-    } else if (props.messages === "sent") {
-      window.getSentMessages().then((messages: Message[]) => {
-        setMessages(messages);
-      });
-    }
-  }, [props.messages]);
+        }
+      );
 
-  const noMessageExplanation =
-    props.messages === "new" ? "No new messages." : "No messages.";
+      return cancel;
+    } else if (type === "all") {
+      setMessages([]);
+      const cancel = window.getMessagesStreamed(
+        { filter: daemon_pb.GetMessagesRequest.Filter.ALL },
+        (messages: IncomingMessage[]) => {
+          setMessages((prev: IncomingMessage[]) => {
+            // merge new messages with old messages, and sort them by timestamp
+            const newMessages = messages.concat(prev);
+            newMessages.sort((a, b) => {
+              // sort based on timestamp
+              const aTime = a.deliveredAt ?? new Date();
+              const bTime = b.deliveredAt ?? new Date();
+              if (aTime > bTime) {
+                return -1;
+              } else if (aTime < bTime) {
+                return 1;
+              } else {
+                return 0;
+              }
+            });
+
+            return newMessages;
+          });
+        }
+      );
+
+      return cancel;
+    }
+
+    // Error out
+    throw new Error("Invalid messages type");
+  }, [type]);
 
   return (
     <div>
-      <div className="flex place-content-center w-full mt-8">
-        <div className="place-self-center flex flex-col w-full max-w-3xl">
+      <div className="mt-8 flex w-full place-content-center">
+        <div className="flex w-full max-w-3xl flex-col place-self-center">
           <SelectableList
             items={messages.map((message) => {
               return {
-                id: message.id,
+                id: message.uid,
                 data: message,
-                action: () => props.readCallback(message),
+                action: () => readCallback(message),
               };
             })}
             searchable={false}
-            globalAction={() => {}}
             onRender={({ item, active }) =>
               typeof item === "string" ? (
                 <div className="unselectable">{item}</div>
               ) : (
-                <MessageBlurb
+                <IncomingMessageBlurb
                   active={active}
                   key={item.id}
                   message={item.data}
@@ -139,11 +179,70 @@ function MessageList(props: {
           />
         </div>
       </div>
-      {messages.length === 0 && (
-        <NoMessages explanation={noMessageExplanation} />
-      )}
+      {messages.length === 0 && <BackgroundCircles />}
     </div>
   );
 }
 
-export default MessageList;
+export function OutgoingMessageList({
+  type,
+  readCallback,
+}: {
+  type: "outbox" | "sent";
+  readCallback: (message: OutgoingMessage) => void;
+}): JSX.Element {
+  const [messages, setMessages] = React.useState<OutgoingMessage[]>([]);
+
+  React.useEffect(() => {
+    if (type === "outbox") {
+      window
+        .getOutboxMessages({})
+        .then(({ messagesList }: { messagesList: OutgoingMessage[] }) => {
+          setMessages(messagesList);
+        })
+        .catch((err) => {
+          console.error(err);
+        });
+    } else if (type === "sent") {
+      window
+        .getSentMessages({})
+        .then(({ messagesList }: { messagesList: OutgoingMessage[] }) => {
+          setMessages(messagesList);
+        })
+        .catch((err) => {
+          console.error(err);
+        });
+    }
+  }, [type]);
+
+  return (
+    <div>
+      <div className="mt-8 flex w-full place-content-center">
+        <div className="flex w-full max-w-3xl flex-col place-self-center">
+          <SelectableList
+            items={messages.map((message) => {
+              return {
+                id: message.uid,
+                data: message,
+                action: () => readCallback(message),
+              };
+            })}
+            searchable={false}
+            onRender={({ item, active }) =>
+              typeof item === "string" ? (
+                <div className="unselectable">{item}</div>
+              ) : (
+                <OutgoingMessageBlurb
+                  active={active}
+                  key={item.id}
+                  message={item.data}
+                />
+              )
+            }
+          />
+        </div>
+      </div>
+      {messages.length === 0 && <BackgroundCircles />}
+    </div>
+  );
+}
