@@ -9,7 +9,7 @@ import { useVirtual } from "react-virtual";
 import { usePointerMovedSinceMount } from "../utils";
 
 export interface ListItem<T> {
-  id: string;
+  id: string | number;
   data: T;
   action: (() => void) | null;
 }
@@ -23,23 +23,28 @@ interface RenderParams<T, TT = ListItem<T> | string> {
 interface SelectableListProps<T, TT = ListItem<T> | string> {
   items: TT[];
   onRender: (params: RenderParams<T>) => React.ReactElement;
-  globalAction: () => void;
+  globalAction?: () => void;
   searchable: boolean;
 }
 
-export function SelectableList<T>(props: SelectableListProps<T>) {
+export function SelectableList<T>({
+  items,
+  onRender,
+  globalAction,
+  searchable,
+}: SelectableListProps<T>): JSX.Element {
   const activeRef = React.useRef<HTMLDivElement>(null);
   const parentRef = React.useRef(null);
 
   let startIndex = 0;
-  while (typeof props.items[startIndex] === "string") startIndex++;
+  while (typeof items[startIndex] === "string") startIndex++;
 
   const [activeIndex, setActiveIndex] = React.useState(startIndex);
 
   // store a ref to all items so we do not have to pass
   // them as a dependency when setting up event listeners.
-  const itemsRef = React.useRef(props.items);
-  itemsRef.current = props.items;
+  const itemsRef = React.useRef(items);
+  itemsRef.current = items;
 
   const rowVirtualizer = useVirtual({
     size: itemsRef.current.length,
@@ -49,41 +54,39 @@ export function SelectableList<T>(props: SelectableListProps<T>) {
   const previousIndex = React.useCallback(
     (oldIndex: number) => {
       let nextIndex = oldIndex > startIndex ? oldIndex - 1 : oldIndex;
-      while (typeof props.items[nextIndex] === "string") {
+      while (typeof items[nextIndex] === "string") {
         if (nextIndex === startIndex) break;
         nextIndex -= 1;
       }
       return nextIndex;
     },
-    [props.items, startIndex]
+    [items, startIndex]
   );
 
   const nextIndex = React.useCallback(
     (oldIndex: number) => {
-      let nextIndex =
-        oldIndex < props.items.length - 1 ? oldIndex + 1 : oldIndex;
-      while (typeof props.items[nextIndex] === "string") {
-        if (nextIndex === props.items.length - 1) break;
+      let nextIndex = oldIndex < items.length - 1 ? oldIndex + 1 : oldIndex;
+      while (typeof items[nextIndex] === "string") {
+        if (nextIndex === items.length - 1) break;
         nextIndex += 1;
       }
       return nextIndex;
     },
-    [props.items]
+    [items]
   );
 
   React.useEffect(() => {
-    if (activeIndex > props.items.length - 1 && props.items.length > 0) {
-      setActiveIndex(props.items.length - 1);
+    if (activeIndex > items.length - 1 && items.length > 0) {
+      setActiveIndex(items.length - 1);
     }
-  }, [props.items.length, activeIndex, setActiveIndex]);
+  }, [items.length, activeIndex, setActiveIndex]);
 
   // Handle keyboard up and down events.
   React.useEffect(() => {
-    const handler = (event: any) => {
-      console.log(`event.keyCode: ${event.key}`);
+    const handler = (event: KeyboardEvent): void => {
       if (
         event.key === "ArrowUp" ||
-        ((event.ctrlKey || !props.searchable) && event.key === "k")
+        ((event.ctrlKey || !searchable) && event.key === "k")
       ) {
         setActiveIndex((old) => {
           console.log("oldIndex: " + old);
@@ -91,7 +94,7 @@ export function SelectableList<T>(props: SelectableListProps<T>) {
         });
       } else if (
         event.key === "ArrowDown" ||
-        ((event.ctrlKey || !props.searchable) && event.key === "j")
+        ((event.ctrlKey || !searchable) && event.key === "j")
       ) {
         setActiveIndex((old) => {
           console.log("oldIndex: " + old);
@@ -109,7 +112,7 @@ export function SelectableList<T>(props: SelectableListProps<T>) {
     };
     window.addEventListener("keydown", handler);
     return () => window.removeEventListener("keydown", handler);
-  }, [setActiveIndex, previousIndex, nextIndex]);
+  }, [setActiveIndex, previousIndex, nextIndex, searchable]);
 
   // destructuring here to prevent linter warning to pass
   // entire rowVirtualizer in the dependencies array.
@@ -123,18 +126,21 @@ export function SelectableList<T>(props: SelectableListProps<T>) {
     });
   }, [activeIndex, scrollToIndex]);
 
-  const execute = React.useCallback((item: RenderParams<T>["item"]) => {
-    if (typeof item === "string") return;
-    if (item.action) {
-      item.action();
-      props.globalAction();
-      // if you click divider, you should only display things in that section
-      // TODO(sualeh): this is a hack, we should be able to do this
-      //   } else {
-      //     query.setSearch("");
-      //     query.setCurrentRootAction(item.id);
-    }
-  }, []);
+  const execute = React.useCallback(
+    (item: RenderParams<T>["item"]) => {
+      if (typeof item === "string") return;
+      if (item.action) {
+        item.action();
+        if (globalAction) globalAction();
+        // if you click divider, you should only display things in that section
+        // TODO(sualeh): this is a hack, we should be able to do this
+        //   } else {
+        //     query.setSearch("");
+        //     query.setCurrentRootAction(item.id);
+      }
+    },
+    [globalAction]
+  );
 
   const pointerMoved = usePointerMovedSinceMount();
 
@@ -143,7 +149,6 @@ export function SelectableList<T>(props: SelectableListProps<T>) {
       <div ref={parentRef} className="relative max-h-full">
         <div
           role="listbox"
-          // TODO(sualeh): ask Arvid about this.
           className="w-full"
           /* This is a hack to make the listbox scrollable. */
           style={{
@@ -151,7 +156,7 @@ export function SelectableList<T>(props: SelectableListProps<T>) {
           }}
         >
           {rowVirtualizer.virtualItems.map((virtualRow) => {
-            const item = itemsRef.current[virtualRow.index];
+            const item = itemsRef.current[virtualRow.index] ?? "";
             const handlers = typeof item !== "string" && {
               onPointerMove: () =>
                 pointerMoved &&
@@ -174,15 +179,13 @@ export function SelectableList<T>(props: SelectableListProps<T>) {
                 }}
                 {...handlers}
               >
-                {React.cloneElement(
-                  props.onRender({
-                    item,
-                    active,
-                  }),
-                  {
-                    ref: virtualRow.measureRef,
-                  }
-                )}
+                {/* This is a hack to make the list resize when a single item changes. See https://github.com/TanStack/virtual/issues/85. 
+                React-virtual v3 has the function virtualizer.measure() which we could call instead, and which would be better than this.
+                When that becomes stable we should use that instead.
+                */}
+                <div ref={(element) => virtualRow.measureRef(element)}>
+                  {onRender({ item, active })}
+                </div>
               </div>
             );
           })}
