@@ -5,6 +5,24 @@
 
 import { render } from "react-dom";
 import Main from "./Main";
+import { InitialModal } from "./Main";
 import "./index.css";
 
-render(<Main />, document.getElementById("root"));
+// get path
+const url = new URL(window.location.href);
+const params = new URLSearchParams(url.search);
+const publicID = params.get("publicID");
+
+let component;
+if (publicID != null) {
+  component = (
+    <Main
+      initialModal={InitialModal.AddFriendByPublicId}
+      initialModalData={publicID}
+    />
+  );
+} else {
+  component = <Main />;
+}
+
+render(component, document.getElementById("root"));
